services/auth/requirements.txt
# Auth Service Dependencies
# Used by Pants for dependency management

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pinned for passlib compatibility (4.1+ breaks passlib)
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
email-validator>=2.1.0

# Development/Debug
debugpy>=1.8.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
